Restructure to lib. Greatly simplify utils.
