+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "since_when"
-version = "0.2.0"
+version = "0.2.1"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -13,3 +13,11 @@ num-traits = "0.2.15"
 [dependencies.rusqlite]
 version = "0.29.0"
 features = ["bundled"]
+
+[[bin]]
+name = "since_when"
+path = "src/main.rs"
+
+[lib]
+name = "since_when_lib"
+path = "src/lib.rs"
