+++ src/lib.rs
@@ -1,0 +1,6 @@
+pub mod add_event;
+pub mod app;
+pub mod calendar;
+pub mod database;
+pub mod events;
+pub mod utils;
